Move Post into separate module
