+++ Cargo.toml
@@ -27,3 +27,4 @@ lazy_static = "1.4.0"
 config = { git = "https://github.com/mehcode/config-rs", branch = "master" }
 chrono = "0.4"
 syntect = "4.6"
+atom_syndication = "0.11"
